feat(memory): remove redundant as i64 casts in miner tests
default_forgetting_initial_ttl_days() already returns i64, so the
5 casts (4 in sqlite_tests.rs, 1 in entity_tests.rs) were unnecessary
after the Step 1.5 type migration.

Source: review-b-followup
Type: cleanup

+++ crates/memory/src/miner_tests/entity_tests.rs
@@ -182,7 +182,7 @@ fn test_per_agent_dedup_same_agent_type_name() {
         "agent-A",
         &events,
         &entities,
-        default_forgetting_initial_ttl_days() as i64,
+        default_forgetting_initial_ttl_days(),
     )
     .unwrap();
     let entity_count: i64 = conn
